backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic
pydantic-settings
openai
anthropic
python-multipart
python-pptx
cryptography
reportlab
python-docx
PyPDF2
jinja2
aiofiles
httpx
requests
